Readability
Increased the readability of the code by adding lots of comments. Also decided to add a saftey check in enqueue to make sure createNode is returning the proper data. Showed in the comments that enqueue is O(1) complexity and dequeue is O(N) complexity. I could eventually make dequeue O(1) complexity by adding a struct that keeps the data for the tail of the queue.